fixed many bugs
