--- packages/client/src/components/Forum/ForumPost/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ForumPost';
