--- packages/client/src/components/Forum/ForumPost/utils/TimeFormatting.tsx
@@ -1,6 +1,0 @@
-export function TimeFormatting(data: string) {
-  const time = new Date(data);
-
-  return `${time.getDay()}.${time.getMonth() + 1}.${time.getFullYear()}
-  ${time.getHours()}:${time.getMinutes()}:${time.getSeconds()}`;
-}
