--- packages/client/src/pages/Forum/pages/Forum/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Forum';
