+++ packages/client/src/pages/Forum/pages/ForumMessages/components/MessageElement/EmojiBox/EmojiBox.tsx
@@ -1,19 +1,19 @@
 import styled from 'styled-components';
 import { H1 } from 'src/design/H1';
-import { EmojiLinkIcon } from 'components/Forum/ForumPost/utils/EmojiLinkIcon';
+import { EmojiLinkIcon } from 'pages/Forum/pages/ForumMessages/utils/emojiLinkIcon';
 
 const EmojiBox = ({ id = 1, num = 1 }) => {
   return (
-    <Containers>
+    <Container>
       <Emogi src={`${EmojiLinkIcon(id)}`} />
       <Number>{num}</Number>
-    </Containers>
+    </Container>
   );
 };
 
 export default EmojiBox;
 
-const Containers = styled.div`
+const Container = styled('div')`
   display: grid;
   grid-template-columns: auto auto;
   grid-column-gap: 10px;
@@ -24,7 +24,7 @@ const Containers = styled.div`
   align-items: center;
 `;
 
-const Emogi = styled.img`
+const Emogi = styled('img')`
   width: 18px;
   height: 18px;
   display: grid;
@@ -39,5 +39,5 @@ const Number = styled(H1)`
   font-weight: 400;
   font-size: 18px;
   line-height: 20px;
-  color: ${({ theme }) => theme.colors.core.text.septenary};
+  color: ${({ theme }) => theme.colors.core.text.primary};
 `;
