+++ packages/client/src/pages/Forum/pages/ForumMessages/demoData.ts
@@ -1,0 +1,132 @@
+export const demoMessage = [
+  {
+    id: 1,
+    data: '2020-01-02T14:22:22.000Z',
+    author: { id: 1, first_name: 'Иван', second_name: 'Karlson' },
+    post_id: 1,
+    message_main_id: 0,
+    emojis: [],
+    text: 'Если нажать два раза прыжок, то взлетишь',
+  },
+  {
+    id: 2,
+    data: '2020-01-02T14:27:55.000Z',
+    author: { id: 2, first_name: 'Митя', second_name: 'Ликупер' },
+    post_id: 1,
+    message_main_id: 1,
+    emojis: [{ id: 1, author_id: [1] }],
+    text: 'Ты все врешь, я только как не нажимал, не чего не получается',
+  },
+  {
+    id: 3,
+    data: '2020-01-02T17:34:04.000Z',
+    author: { id: 1, first_name: 'Иван', second_name: 'Karlson' },
+    post_id: 1,
+    message_main_id: 0,
+    emojis: [],
+    text: 'Если нажать два раза прыжок, то взлетишь',
+  },
+  {
+    id: 4,
+    data: '2020-01-02T20:01:20.000Z',
+    author: { id: 1, first_name: 'Ivan', second_name: 'Karlson' },
+    post_id: 1,
+    message_main_id: 1,
+    emojis: [
+      { id: 1, author_id: [1, 2] },
+      { id: 2, author_id: [1] },
+    ],
+    text: 'Значит ты рукожоп, у меня все норм',
+  },
+  {
+    id: 5,
+    data: '2020-01-02T20:01:20.000Z',
+    author: { id: 1, first_name: 'Ivan', second_name: 'Karlson' },
+    post_id: 1,
+    message_main_id: 1,
+    emojis: [
+      { id: 1, author_id: [1, 2] },
+      { id: 2, author_id: [1] },
+    ],
+    text: 'Значит ты рукожоп, у меня все норм',
+  },
+  {
+    id: 6,
+    data: '2020-01-02T20:01:20.000Z',
+    author: { id: 1, first_name: 'Ivan', second_name: 'Karlson' },
+    post_id: 1,
+    message_main_id: 1,
+    emojis: [
+      { id: 1, author_id: [1, 2] },
+      { id: 2, author_id: [1] },
+    ],
+    text: 'Значит ты рукожоп, у меня все норм',
+  },
+  {
+    id: 7,
+    data: '2020-01-02T20:01:20.000Z',
+    author: { id: 1, first_name: 'Ivan', second_name: 'Karlson' },
+    post_id: 1,
+    message_main_id: 1,
+    emojis: [
+      { id: 1, author_id: [1, 2] },
+      { id: 2, author_id: [1] },
+    ],
+    text: 'Значит ты рукожоп, у меня все норм',
+  },
+  {
+    id: 8,
+    data: '2020-01-02T20:01:20.000Z',
+    author: { id: 1, first_name: 'Ivan', second_name: 'Karlson' },
+    post_id: 1,
+    message_main_id: 1,
+    emojis: [
+      { id: 1, author_id: [1, 2] },
+      { id: 2, author_id: [1] },
+    ],
+    text: 'Значит ты рукожоп, у меня все норм',
+  },
+  {
+    id: 9,
+    data: '2020-01-02T20:01:20.000Z',
+    author: { id: 1, first_name: 'Ivan', second_name: 'Karlson' },
+    post_id: 1,
+    message_main_id: 1,
+    emojis: [
+      { id: 1, author_id: [1, 2] },
+      { id: 2, author_id: [1] },
+    ],
+    text: 'Значит ты рукожоп, у меня все норм',
+  },
+  {
+    id: 10,
+    data: '2020-01-02T20:01:20.000Z',
+    author: { id: 1, first_name: 'Ivan', second_name: 'Karlson' },
+    post_id: 1,
+    message_main_id: 1,
+    emojis: [
+      { id: 1, author_id: [1, 2] },
+      { id: 2, author_id: [1] },
+    ],
+    text: 'Значит ты рукожоп, у меня все норм',
+  },
+  {
+    id: 11,
+    data: '2020-01-02T20:01:20.000Z',
+    author: { id: 1, first_name: 'Ivan', second_name: 'Karlson' },
+    post_id: 1,
+    message_main_id: 1,
+    emojis: [
+      { id: 1, author_id: [1, 2] },
+      { id: 2, author_id: [1] },
+    ],
+    text: 'Значит ты рукожоп, у меня все норм',
+  },
+];
+
+export const demoPost = {
+  id: 1,
+  data: '2020-01-02T14:22:22.000Z',
+  author: { id: 1, first_name: 'Ivan', second_name: 'Karlson' },
+  title: 'Это новый лайвхак. Как обойти правила в этой игре, и все выиграть',
+};
