+++ packages/client/src/pages/Forum/pages/ForumMessages/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ForumMessages';
