+++ packages/client/src/pages/Forum/pages/ForumMessages/utils/emojiLinkIcon.tsx
