+++ packages/client/src/pages/Forum/pages/ForumPosts/components/ForumPostsInput/index.ts
