--- packages/client/src/pages/Forum/pages/Post/Post.tsx
@@ -1,20 +1,0 @@
-import styled from 'styled-components';
-import ForumPost from 'components/Forum/ForumPost';
-
-const Post = () => {
-  return (
-    <Container>
-      <ForumPost />
-    </Container>
-  );
-};
-
-export default Post;
-
-export const Container = styled.div`
-  height: 100vh;
-  width: 100vw;
-  display: grid;
-  align-items: center;
-  justify-items: center;
-`;
