--- packages/client/src/pages/Forum/pages/Post/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Post';
